+++ pkg/ffuf/autocalibration.go
@@ -1,14 +1,14 @@
 package ffuf
 
 import (
+	"encoding/json"
 	"fmt"
 	"log"
 	"math/rand"
+	"os"
+	"path/filepath"
 	"strconv"
 	"time"
-	"encoding/json"
-	"path/filepath"
-	"os"
 )
 
 type AutocalibrationStrategy map[string][]string
@@ -20,9 +20,9 @@ func (j *Job) autoCalibrationStrings() map[string][]string {
 	if len(j.Config.AutoCalibrationStrings) > 0 {
 		cInputs["custom"] = append(cInputs["custom"], j.Config.AutoCalibrationStrings...)
 		return cInputs
-		
+
 	}
-		
+
 	for _, strategy := range j.Config.AutoCalibrationStrategies {
 		jsonStrategy, err := os.ReadFile(filepath.Join(AUTOCALIBDIR, strategy+".json"))
 		if err != nil {
@@ -36,36 +36,36 @@ func (j *Job) autoCalibrationStrings() map[string][]string {
 			j.Output.Warning(fmt.Sprintf("Skipping strategy \"%s\" because of error: %s\n", strategy, err))
 			continue
 		}
-		
+
 		cInputs = mergeMaps(cInputs, tmpStrategy)
 	}
-	
+
 	return cInputs
 }
 
 func setupDefaultAutocalibrationStrategies() error {
-	basic_strategy := AutocalibrationStrategy {
-		"basic_admin": []string{"admin"+RandomString(16), "admin"+RandomString(8)},
-		"htaccess": []string{".htaccess"+RandomString(16), ".htaccess"+RandomString(8)},
+	basic_strategy := AutocalibrationStrategy{
+		"basic_admin":  []string{"admin" + RandomString(16), "admin" + RandomString(8)},
+		"htaccess":     []string{".htaccess" + RandomString(16), ".htaccess" + RandomString(8)},
 		"basic_random": []string{RandomString(16), RandomString(8)},
 	}
 	basic_strategy_json, err := json.Marshal(basic_strategy)
 	if err != nil {
 		return err
 	}
-	
-	advanced_strategy := AutocalibrationStrategy {
-		"basic_admin": []string{"admin"+RandomString(16), "admin"+RandomString(8)},
-		"htaccess": []string{".htaccess"+RandomString(16), ".htaccess"+RandomString(8)},
+
+	advanced_strategy := AutocalibrationStrategy{
+		"basic_admin":  []string{"admin" + RandomString(16), "admin" + RandomString(8)},
+		"htaccess":     []string{".htaccess" + RandomString(16), ".htaccess" + RandomString(8)},
 		"basic_random": []string{RandomString(16), RandomString(8)},
-		"admin_dir": []string{"admin"+RandomString(16)+"/", "admin"+RandomString(8)+"/"},
-		"random_dir": []string{RandomString(16)+"/", RandomString(8)+"/"},
+		"admin_dir":    []string{"admin" + RandomString(16) + "/", "admin" + RandomString(8) + "/"},
+		"random_dir":   []string{RandomString(16) + "/", RandomString(8) + "/"},
 	}
 	advanced_strategy_json, err := json.Marshal(advanced_strategy)
 	if err != nil {
 		return err
 	}
-	
+
 	basic_strategy_file := filepath.Join(AUTOCALIBDIR, "basic.json")
 	if !FileExists(basic_strategy_file) {
 		err = os.WriteFile(filepath.Join(AUTOCALIBDIR, "basic.json"), basic_strategy_json, 0640)
@@ -76,7 +76,7 @@ func setupDefaultAutocalibrationStrategies() error {
 		err = os.WriteFile(filepath.Join(AUTOCALIBDIR, "advanced.json"), advanced_strategy_json, 0640)
 		return err
 	}
-	
+
 	return nil
 }